Change span fields to usize

+++ lily-ast/src/ann.rs
@@ -2,7 +2,7 @@
 
 /// A pair of starting and ending byte offsets.
 #[derive(Debug, PartialEq, Eq, Hash)]
-pub struct Span(pub u32, pub u32);
+pub struct Span(pub usize, pub usize);
 
 /// The provenance of an AST node.
 #[derive(Debug, PartialEq, Eq, Hash)]
